cor.mtest {corrplot}	R Documentation

Significance test which produces p-values and confidence intervals for each pair of input features.

Description

Significance test which produces p-values and confidence intervals for each pair of input features.

Usage

cor.mtest(mat, ...)

Arguments

mat
Input matrix of size NxF, with N rows that represent samples and F columns that represent features.

...
Additional arguments passed to function cor.test, e.g. conf.level = 0.95.

Value

Return a list containing:

p
Square matrix of size FxF with p-values as cells

lowCI
Square matrix of size FxF, each cell represents the lower part of a confidence interval

uppCI
Square matrix of size FxF, each cell represents the upper part of a confidence interval

See Also

Function cor.test

[Package corrplot version 0.92 Index]